pyproject.toml
[tool.poetry]
name = "steradian"
version = "0.0.1"
description = "Cars rental"
authors = ["Andrie Gunawan <andriebamz@gmail.com>"]
readme = "README.md"
packages = [
    {include = "core"},
    {include = "api"},
]

[tool.poetry.scripts]
steradian = "api:main"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.1"
django-filter = "^23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
